Match set copying behavior
Original deepcopy doesn't have specialized handler for sets, instead
it uses __reduce__ protocol, so it never iterates over original set.

Arguably this is redundant, but if set changes during iteration,
we would get an error with previous approach as opposed to stdlib.

Maybe it's worth to revert it at some point.